[auth] Add hook for refreshing the user store
