+++ src/client/hooks/auto-login.hook.ts
@@ -1,8 +1,7 @@
-import { useAppDispatch, useAppSelector } from "@hooks/store-hook";
-import { selectAuthenticatedUser, setUser } from "@stores/user.store";
-import { ApiResponse } from "@typing/api-response.interface";
+import { useRefreshUser } from "@hooks/refresh-user.hook";
+import { useAppSelector } from "@hooks/store-hook";
+import { selectAuthenticatedUser } from "@stores/user.store";
 import { IUserPublic } from "@typing/user.interface";
-import axios, { AxiosResponse } from "axios";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
 
@@ -10,7 +9,7 @@ export const useAutoLogin = (): IUserPublic | undefined => {
   const currentUser: IUserPublic | undefined = useAppSelector(
     selectAuthenticatedUser
   );
-  const dispatch = useAppDispatch();
+  const refreshUser = useRefreshUser();
   const router = useRouter();
 
   useEffect(() => {
@@ -23,24 +22,8 @@ export const useAutoLogin = (): IUserPublic | undefined => {
     }
 
     // Else, get current session
-    axios
-      .get<ApiResponse<{ user: IUserPublic }>>("/api/user/login")
-      .then(
-        ({
-          data: response,
-        }: AxiosResponse<ApiResponse<{ user: IUserPublic }>>) => {
-          if (response.success) {
-            console.log(
-              `Auto logging de l'utilisateur ${response.data?.user.email}`
-            );
-            dispatch(setUser(response.data?.user));
-          } else {
-            console.warn("La session a expirée");
-          }
-        }
-      )
-      .catch(() => console.warn("Erreur autologin"));
-  }, [currentUser, dispatch, router]);
+    refreshUser();
+  }, [currentUser, refreshUser, router]);
 
   return currentUser;
 };
